Add CppTest Unit Tests
