+++ CMakeLists.txt
@@ -2,17 +2,10 @@ cmake_minimum_required(VERSION 2.8)
 project(ibmiotf-cpp)
 enable_testing()
 
-#SET(CMAKE_CXX_FLAGS "-g -O0 -Wall -fprofile-arcs -ftest-coverage -fPIC ")
-#SET(CMAKE_C_FLAGS "-g -O0 -Wall -W -fprofile-arcs -ftest-coverage -fPIC ")
-#SET(CMAKE_EXE_LINKER_FLAGS "-fprofile-arcs -ftest-coverage ")
 option (run_tests "set run_tests to ON if build tests should be run, set to OFF to skip tests" ON)
-SET(CMAKE_CXX_FLAGS "-g -O0 -Wall -fprofile-arcs -ftest-coverage -fPIC -std=c++11 -pthread ${CMAKE_CXX_FLAGS} ")
+SET(CMAKE_CXX_FLAGS "-g -O0 -Wall -fprofile-arcs -ftest-coverage -fPIC -std=c++11 -pthread ${CMAKE_CXX_FLAGS} -I/usr/local/include ")
 SET(CMAKE_C_FLAGS "-g -O0 -Wall -W -fprofile-arcs -ftest-coverage -fPIC ${CMAKE_C_FLAGS} ")
-SET(CMAKE_EXE_LINKER_FLAGS "-fprofile-arcs -ftest-coverage ${CMAKE_EXE_LINKER_FLAGS} ")
-
-#if (run_tests)
-#    add_subdirectory(test)
-#endif ()
+SET(CMAKE_EXE_LINKER_FLAGS "-fprofile-arcs -ftest-coverage ${CMAKE_EXE_LINKER_FLAGS} -L/usr/local/lib ")
 
 add_definitions(-DOPENSSL)
 SET(OPENSSL_SEARCH_PATH "" CACHE PATH "Directory containing OpenSSL libraries and includes")
@@ -56,3 +49,7 @@ SET ( LOG4CPP_LIBRARY_NAME "log4cpp" )
 add_subdirectory(src)
 add_subdirectory(lib)
 add_subdirectory(samples)
+
+if (run_tests)
+    add_subdirectory(test)
+endif ()
