+++ samples/CMakeLists.txt
@@ -9,15 +9,14 @@ set(LIBS
     IOTP_ResponseHandler
     )
 
-add_executable(sslDeviceSample sslDeviceSample.cpp)
-add_executable(sslGatewaySample sslGatewaySample.cpp)
 add_executable(sampleQuickstart sampleQuickstart.cpp)
 add_executable(sampleDevice sampleDevice.cpp)
 add_executable(sampleGateway sampleGateway.cpp)
-
-target_link_libraries(sslDeviceSample  IOTP_DeviceClient)
-target_link_libraries(sslGatewaySample  IOTP_GatewayClient)
+add_executable(sampleManagedDevice sampleManagedDevice.cpp)
+add_executable(sampleManagedGateway sampleManagedGateway.cpp)
 
 target_link_libraries(sampleQuickstart  IOTP_DeviceClient IOTP_ResponseHandler)
-target_link_libraries(sampleGateway  IOTP_GatewayClient ${LIBS})
-target_link_libraries(sampleDevice  IOTP_DeviceClient ${LIBS})
+target_link_libraries(sampleDevice  IOTP_DeviceClient)
+target_link_libraries(sampleGateway  IOTP_GatewayClient)
+target_link_libraries(sampleManagedGateway  IOTP_GatewayClient ${LIBS})
+target_link_libraries(sampleManagedDevice  IOTP_DeviceClient ${LIBS})
