+++ test/gateway.cfg
@@ -1,0 +1,9 @@
+{
+	"Organization-ID": $orgId,
+	"domain": $domain,
+	"Device-Type": $deviceType,
+	"Device-ID": $deviceId,
+	"Authentication-Method": $authMethod,
+	"Authentication-Token": $authToken,
+	"clientTrustStorePath" : "iot-cpp-home/IoTFoundation.pem"
+}
