+++ test/test.cfg
@@ -1,0 +1,9 @@
+{
+	"Organization-ID": "testorg",
+	"Domain": "internetofthings.ibmcloud.com",
+	"Device-Type": "type",
+	"Device-ID": "id",
+	"Authentication-Method": "token",
+	"Authentication-Token": "password",
+	"clientTrustStorePath" : "IoTFoundation.pem"
+}
